@@ -1,17 +1,14 @@
 ﻿using Blog.Core.Common.LogHelper;
 using Blog.Core.Hubs;
-using Blog.Core.Log;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.SignalR;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 using StackExchange.Profiling;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading.Tasks;
 
 namespace Blog.Core.Filter
 {
@@ -20,11 +17,14 @@ namespace Blog.Core.Filter
     /// </summary>
     public class GlobalExceptionsFilter : IExceptionFilter
     {
-        private readonly IHostingEnvironment _env;
-        private readonly ILoggerHelper _loggerHelper;
+        private readonly IWebHostEnvironment _env;
         private readonly IHubContext<ChatHub> _hubContext;
+        private readonly ILogger<GlobalExceptionsFilter> _loggerHelper;
+        private static readonly log4net.ILog log =
+        log4net.LogManager.GetLogger(typeof(GlobalExceptionsFilter));
 
-        public GlobalExceptionsFilter(IHostingEnvironment env, ILoggerHelper loggerHelper, IHubContext<ChatHub> hubContext)
+
+        public GlobalExceptionsFilter(IWebHostEnvironment env, ILogger<GlobalExceptionsFilter> loggerHelper, IHubContext<ChatHub> hubContext)
         {
             _env = env;
             _loggerHelper = loggerHelper;
@@ -46,7 +46,7 @@ namespace Blog.Core.Filter
 
 
             //采用log4net 进行错误日志记录
-            _loggerHelper.Error(json.Message, WriteLog(json.Message, context.Exception));
+            log.Error(json.Message + WriteLog(json.Message, context.Exception));
 
             _hubContext.Clients.All.SendAsync("ReceiveUpdate", LogLock.GetLogData()).Wait();
 
@@ -60,7 +60,7 @@ namespace Blog.Core.Filter
         /// <returns></returns>
         public string WriteLog(string throwMsg, Exception ex)
         {
-            return string.Format("【自定义错误】：{0} \r\n【异常类型】：{1} \r\n【异常信息】：{2} \r\n【堆栈调用】：{3}", new object[] { throwMsg,
+            return string.Format("\r\n【自定义错误】：{0} \r\n【异常类型】：{1} \r\n【异常信息】：{2} \r\n【堆栈调用】：{3}", new object[] { throwMsg,
                 ex.GetType().Name, ex.Message, ex.StackTrace });
         }
 
