@@ -17,10 +17,11 @@ namespace Blog.Core.Common.LogHelper
         static ReaderWriterLockSlim LogWriteLock = new ReaderWriterLockSlim();
         static int WritedCount = 0;
         static int FailedCount = 0;
-        static string contentRoot = string.Empty;
+        static string _contentRoot = string.Empty;
 
-        public LogLock(string ContentRootPath) {
-            contentRoot = ContentRootPath;
+        public LogLock(string contentPath)
+        {
+            _contentRoot = contentPath;
         }
 
         public static void OutSql2Log(string filename, string[] dataParas)
@@ -33,7 +34,7 @@ namespace Blog.Core.Common.LogHelper
                 //      因进入与退出写入模式应在同一个try finally语句块内，所以在请求进入写入模式之前不能触发异常，否则释放次数大于请求次数将会触发异常
                 LogWriteLock.EnterWriteLock();
 
-                var path = Path.Combine(contentRoot, "Log");
+                var path = Path.Combine(_contentRoot, "Log");
                 if (!Directory.Exists(path))
                 {
                     Directory.CreateDirectory(path);
@@ -104,7 +105,7 @@ namespace Blog.Core.Common.LogHelper
             List<LogInfo> reqresLogs = new List<LogInfo>();
             try
             {
-                aopLogs = ReadLog(Path.Combine(contentRoot, "Log", "AOPLog.log"), Encoding.UTF8)
+                aopLogs = ReadLog(Path.Combine(_contentRoot, "Log", "AOPLog.log"), Encoding.UTF8)
                 .Split("--------------------------------")
                 .Where(d => !string.IsNullOrEmpty(d) && d != "\n" && d != "\r\n")
                 .Select(d => new LogInfo
@@ -121,7 +122,7 @@ namespace Blog.Core.Common.LogHelper
 
             try
             {
-                excLogs = ReadLog(Path.Combine(contentRoot, "Log", $"GlobalExcepLogs_{DateTime.Now.ToString("yyyMMdd")}.log"), Encoding.UTF8)?
+                excLogs = ReadLog(Path.Combine(_contentRoot, "Log", $"GlobalExcepLogs_{DateTime.Now.ToString("yyyMMdd")}.log"), Encoding.UTF8)?
                       .Split("--------------------------------")
                       .Where(d => !string.IsNullOrEmpty(d) && d != "\n" && d != "\r\n")
                       .Select(d => new LogInfo
@@ -139,7 +140,7 @@ namespace Blog.Core.Common.LogHelper
 
             try
             {
-                sqlLogs = ReadLog(Path.Combine(contentRoot, "Log", "SqlLog.log"), Encoding.UTF8)
+                sqlLogs = ReadLog(Path.Combine(_contentRoot, "Log", "SqlLog.log"), Encoding.UTF8)
                       .Split("--------------------------------")
                       .Where(d => !string.IsNullOrEmpty(d) && d != "\n" && d != "\r\n")
                       .Select(d => new LogInfo
@@ -155,7 +156,7 @@ namespace Blog.Core.Common.LogHelper
 
             try
             {
-                reqresLogs = ReadLog(Path.Combine(contentRoot, "Log", "RequestResponseLog.log"), Encoding.UTF8)
+                reqresLogs = ReadLog(Path.Combine(_contentRoot, "Log", "RequestResponseLog.log"), Encoding.UTF8)
                       .Split("--------------------------------")
                       .Where(d => !string.IsNullOrEmpty(d) && d != "\n" && d != "\r\n")
                       .Select(d => new LogInfo
