@@ -1,14 +1,12 @@
 ﻿using Blog.Core.IServices.BASE;
 using Blog.Core.Model.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Blog.Core.IServices
 {
     public partial interface IGuestbookServices : IBaseServices<Guestbook>
     {
+        Task<bool> TestTranInRepository();
+        Task<bool> TestTranInRepositoryAOP();
     }
 }
