@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Blog.Core.IRepository;
+﻿using Blog.Core.IRepository;
+using Blog.Core.IRepository.UnitOfWork;
 using Blog.Core.Model.Models;
 using Blog.Core.Repository.Base;
 
@@ -11,6 +7,8 @@ namespace Blog.Core.Repository
 {
     public partial class PasswordLibRepository : BaseRepository<PasswordLib>, IPasswordLibRepository
     {
-
+        public PasswordLibRepository(IUnitOfWork unitOfWork) : base(unitOfWork)
+        {
+        }
     }
 }
