@@ -1,15 +1,17 @@
 using Blog.Core.IRepository;
+using Blog.Core.IRepository.UnitOfWork;
 using Blog.Core.Model.Models;
 using Blog.Core.Repository.Base;
 
 namespace Blog.Core.Repository
-{	
-	/// <summary>
-	/// sysUserInfoRepository
-	/// </summary>	
-	public class sysUserInfoRepository : BaseRepository<sysUserInfo>, IsysUserInfoRepository
+{
+    /// <summary>
+    /// sysUserInfoRepository
+    /// </summary>	
+    public class sysUserInfoRepository : BaseRepository<sysUserInfo>, IsysUserInfoRepository
     {
-
-       
+        public sysUserInfoRepository(IUnitOfWork unitOfWork) : base(unitOfWork)
+        {
+        }
     }
 }
