@@ -58,12 +58,14 @@ namespace Blog.Core.FrameWork.Services
         {
             string roleName = "";
             var user = (await base.Query(a => a.uLoginName == loginName && a.uLoginPWD == loginPwd)).FirstOrDefault();
+            var roleList = await _roleRepository.Query(a => a.IsDeleted==false);
             if (user != null)
             {
                 var userRoles = await _userRoleServices.Query(ur => ur.UserId == user.uID);
                 if (userRoles.Count > 0)
                 {
-                    var roles = await _roleRepository.QueryByIDs(userRoles.Select(ur => ur.RoleId.ObjToString()).ToArray());
+                    var arr = userRoles.Select(ur => ur.RoleId.ObjToString()).ToList();
+                    var roles = roleList.Where(d => arr.Contains(d.Id.ObjToString()));
 
                     roleName = string.Join(',', roles.Select(r => r.Name).ToArray());
                 }
