@@ -3,18 +3,16 @@ using Autofac.Extensions.DependencyInjection;
 using Autofac.Extras.DynamicProxy;
 using AutoMapper;
 using Blog.Core.Common;
+using Blog.Core.Common.DB;
 using Blog.Core.IServices;
 using Blog.Core.Model.Models;
 using Blog.Core.Services;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Hosting.Internal;
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
+using Microsoft.Extensions.Hosting.Internal;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using Xunit;
 
 namespace Blog.Core.Tests
@@ -30,7 +28,17 @@ namespace Blog.Core.Tests
             IServiceCollection services = new ServiceCollection();
             services.AddAutoMapper(typeof(Startup));
 
-            services.AddSingleton<IHostingEnvironment, HostingEnvironment>();
+            services.AddScoped<SqlSugar.ISqlSugarClient>(o =>
+            {
+                return new SqlSugar.SqlSugarClient(new SqlSugar.ConnectionConfig()
+                {
+                    ConnectionString = BaseDBConfig.ConnectionString,//必填, 数据库连接字符串
+                    DbType = (SqlSugar.DbType)BaseDBConfig.DbType,//必填, 数据库类型
+                    IsAutoCloseConnection = true,//默认false, 时候知道关闭数据库连接, 设置为true无需使用using或者Close操作
+                    IsShardSameThread = true,//共享线程
+                    InitKeyType = SqlSugar.InitKeyType.SystemTable//默认SystemTable, 字段信息读取, 如：该属性是不是主键，标识列等等信息
+                });
+            });
 
             //services.AddSingleton(new Appsettings(Env));
 
@@ -75,12 +83,22 @@ namespace Blog.Core.Tests
             IServiceCollection services = new ServiceCollection();
             services.AddAutoMapper(typeof(Startup));
 
-            services.AddScoped<IHostingEnvironment, HostingEnvironment>();
             services.AddSingleton(new Appsettings(basePath));
             services.AddSingleton<IRedisCacheManager, RedisCacheManager>();
             services.AddScoped<Blog.Core.Model.Models.DBSeed>();
             services.AddScoped<Blog.Core.Model.Models.MyContext>();
 
+            services.AddScoped<SqlSugar.ISqlSugarClient>(o =>
+            {
+                return new SqlSugar.SqlSugarClient(new SqlSugar.ConnectionConfig()
+                {
+                    ConnectionString = BaseDBConfig.ConnectionString,//必填, 数据库连接字符串
+                    DbType = (SqlSugar.DbType)BaseDBConfig.DbType,//必填, 数据库类型
+                    IsAutoCloseConnection = true,//默认false, 时候知道关闭数据库连接, 设置为true无需使用using或者Close操作
+                    IsShardSameThread = true,//共享线程
+                    InitKeyType = SqlSugar.InitKeyType.SystemTable//默认SystemTable, 字段信息读取, 如：该属性是不是主键，标识列等等信息
+                });
+            });
 
             //实例化 AutoFac  容器   
             var builder = new ContainerBuilder();
