@@ -101,7 +101,7 @@ namespace Blog.Core.AOP
         private async Task TestActionAsync(IInvocation invocation)
         {
             //Console.WriteLine("Waiting after method execution for " + invocation.MethodInvocationTarget.Name);
-            //await Task.Delay(200); // 仅作测试
+            await Task.Delay(20); // 仅作测试
             //Console.WriteLine("Waited after method execution for " + invocation.MethodInvocationTarget.Name);
         }
 
