@@ -22,6 +22,8 @@ namespace Blog.Core.AuthHelper
         /// 验证方案提供对象
         /// </summary>
         public IAuthenticationSchemeProvider Schemes { get; set; }
+        private readonly IHttpContextAccessor _accessor;
+
 
         /// <summary>
         /// services 层注入
@@ -33,8 +35,10 @@ namespace Blog.Core.AuthHelper
         /// </summary>
         /// <param name="schemes"></param>
         /// <param name="roleModulePermissionServices"></param>
-        public PermissionHandler(IAuthenticationSchemeProvider schemes, IRoleModulePermissionServices roleModulePermissionServices)
+        /// <param name="accessor"></param>
+        public PermissionHandler(IAuthenticationSchemeProvider schemes, IRoleModulePermissionServices roleModulePermissionServices, IHttpContextAccessor accessor)
         {
+            _accessor = accessor;
             Schemes = schemes;
             this.RoleModulePermissionServices = roleModulePermissionServices;
         }
@@ -42,6 +46,20 @@ namespace Blog.Core.AuthHelper
         // 重写异步处理程序
         protected override async Task HandleRequirementAsync(AuthorizationHandlerContext context, PermissionRequirement requirement)
         {
+            /*
+             * .netcore3.0 启用EndpointRouting后，权限filter不再添加到ActionDescriptor ，而将权限直接作为中间件运行，
+             * 同时所有filter都会添加到endpoint.Metadata。因此，文中的
+             * context.Resource as Microsoft.AspNetCore.Mvc.Filters.AuthorizationFilterContext不再成立。
+             * 
+             * 解决方案有两个：
+             * 
+             * 首先必须在 controller 上进行配置 Authorize ，可以策略授权，也可以角色等基本授权
+             * 
+             * 1、开启公约， startup 中的全局授权过滤公约：o.Conventions.Insert(0, new GlobalRouteAuthorizeConvention());
+             * 
+             * 2、不开启公约，使用 IHttpContextAccessor ，也能实现效果，但是不能自定义返回格式，详细看下边配置；
+             */
+
             // 将最新的角色和接口列表更新
             var data = await RoleModulePermissionServices.GetRoleModule();
             var list = (from item in data
@@ -59,6 +77,12 @@ namespace Blog.Core.AuthHelper
             //从AuthorizationHandlerContext转成HttpContext，以便取出表求信息
             var filterContext = (context.Resource as Microsoft.AspNetCore.Mvc.Filters.AuthorizationFilterContext);
             var httpContext = (context.Resource as Microsoft.AspNetCore.Mvc.Filters.AuthorizationFilterContext)?.HttpContext;
+
+            if (httpContext == null)
+            {
+                httpContext = _accessor.HttpContext;
+            }
+
             //请求Url
             if (httpContext != null)
             {
@@ -69,8 +93,8 @@ namespace Blog.Core.AuthHelper
                 {
                     if (await handlers.GetHandlerAsync(httpContext, scheme.Name) is IAuthenticationRequestHandler handler && await handler.HandleRequestAsync())
                     {
-                        //context.Fail();
-                        //return;
+                        context.Fail();
+                        return;
 
 
                         //自定义返回数据
@@ -93,23 +117,7 @@ namespace Blog.Core.AuthHelper
 
                         httpContext.User = result.Principal;
 
-                        // 取消对URL的判断，因为只需判断该角色下是否匹配当前URL即可，若不匹配都是无效请求
-                        //var isMatchUrl = false;
-                        //var permisssionGroup = requirement.Permissions.GroupBy(g => g.Url);
-                        //foreach (var item in permisssionGroup)
-                        //{
-                        //    try
-                        //    {
-                        //        if (Regex.Match(questUrl, item.Key?.ObjToString().ToLower())?.Value == questUrl)
-                        //        {
-                        //            isMatchUrl = true;
-                        //            break;
-                        //        }
-                        //    }
-                        //    catch (Exception)
-                        //    {
-                        //    }
-                        //}
+
 
                         //权限中是否存在请求的url
                         //if (requirement.Permissions.GroupBy(g => g.Url).Where(w => w.Key?.ToLower() == questUrl).Count() > 0)
@@ -146,18 +154,17 @@ namespace Blog.Core.AuthHelper
 
 
                                 // 可以在这里设置跳转页面
-                                //httpContext.Response.Redirect(requirement.DeniedAction);
-                                //context.Succeed(requirement);
-                                //return;
+                                context.Fail();
+                                return;
 
 
 
                                 //自定义返回数据
-                                var payload = JsonConvert.SerializeObject(new { Code = "403", Message = "很抱歉，您无权访问该接口!" });
-                                httpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
-                                filterContext.Result = new JsonResult(payload);
-                                context.Succeed(requirement);
-                                return;
+                                //var payload = JsonConvert.SerializeObject(new { Code = "403", Message = "很抱歉，您无权访问该接口!" });
+                                //httpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
+                                ////filterContext.Result = new JsonResult(payload);
+                                //context.Succeed(requirement);
+                                //return;
 
 
                             }
@@ -175,16 +182,16 @@ namespace Blog.Core.AuthHelper
                         }
                         else
                         {
-                            //context.Fail();
-                            //return;
+                            context.Fail();
+                            return;
 
 
                             //自定义返回数据
-                            var payload = JsonConvert.SerializeObject(new { Code = "401", Message = "很抱歉，您无权访问该接口!" });
-                            httpContext.Response.StatusCode = StatusCodes.Status401Unauthorized;
-                            filterContext.Result = new JsonResult(payload);
-                            context.Succeed(requirement);
-                            return;
+                            //var payload = JsonConvert.SerializeObject(new { Code = "401", Message = "很抱歉，您无权访问该接口!" });
+                            //httpContext.Response.StatusCode = StatusCodes.Status401Unauthorized;
+                            //filterContext.Result = new JsonResult(payload);
+                            //context.Succeed(requirement);
+                            //return;
                         }
                         return;
                     }
@@ -193,14 +200,16 @@ namespace Blog.Core.AuthHelper
                 if (!questUrl.Equals(requirement.LoginPath.ToLower(), StringComparison.Ordinal) && (!httpContext.Request.Method.Equals("POST")
                                                                                                     || !httpContext.Request.HasFormContentType))
                 {
-                    //context.Fail();
-                    //return;
+                    context.Fail();
+                    return;
 
 
                     //自定义返回数据
-                    var payload = JsonConvert.SerializeObject(new { Code = "401", Message = "很抱歉，您无权访问该接口!" });
-                    httpContext.Response.StatusCode = StatusCodes.Status401Unauthorized;
-                    filterContext.Result = new JsonResult(payload);
+                    //var payload = JsonConvert.SerializeObject(new { Code = "401", Message = "很抱歉，您无权访问该接口!" });
+                    //httpContext.Response.StatusCode = StatusCodes.Status401Unauthorized;
+                    ////filterContext.Result = new JsonResult(payload);
+                    //context.Succeed(requirement);
+                    //return;
                 }
             }
 
