@@ -1,16 +1,6 @@
-﻿using Blog.Core.Common.LogHelper;
-using Blog.Core.Hubs;
-using Blog.Core.Log;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.ApplicationModels;
+﻿using Microsoft.AspNetCore.Mvc.ApplicationModels;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.AspNetCore.Mvc.Filters;
-using Microsoft.AspNetCore.SignalR;
-using StackExchange.Profiling;
-using System;
 using System.Linq;
 using System.Threading.Tasks;
 
